make run_importer_job_03_check_dpae now works locally \o/
